requirements.txt
fast-hotels>=0.2.0
googlemaps>=4.10.0
rich>=13.0.0
python-dotenv>=1.0.0
Jinja2>=3.1.0
